feat(optimizer): extract lowering optimizer from interpreter issue #226
Move optimize_program, optimize_nth_style_recursion, is_tail_recursive_step,
and recursive_call_count from interpreter.py into a dedicated optimizer.py.
All four names re-exported from interpreter.py for backward compatibility.
No behavioral changes.
